web/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.taskipy.tasks]
run-server="poetry run uvicorn server.app:app --reload"
env="poetry run python ./server/settings/utils/environments.py"
aerich-init="poetry run aerich init -t server.settings.database.TORTOISE_ORM --location ./server/models/migrations"
init-db="poetry run aerich init-db"
makemigration="poetry run aerich migrate"
migrate="poetry run aerich upgrade"
downgrade="poetry run aerich downgrade"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
tortoise = "^0.1.1"
taskipy = "^1.13.0"
uvicorn = "^0.31.1"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
tortoise-orm = "^0.21.6"
pydantic = "^2.9.2"
aerich = "^0.7.2"

[tool.aerich]
tortoise_orm = "server.settings.database.TORTOISE_ORM"
location = "./server/models/migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
